requirements.txt
pywin32>=305
keyboard>=0.13.5
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
psutil>=5.9.0
